+++ packages/gallery/src/components/item/imageItem.js
@@ -131,7 +131,6 @@ export default class ImageItem extends GalleryComponent {
               ...restOfDimensions,
               backgroundSize: '0.3px',
               backgroundRepeat: 'repeat',
-              transition: 'all 3s ease-in-out'
             };
             preload = <img
               alt=''