+++ packages/gallery/src/components/item/itemHover.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { utils } from 'pro-gallery-lib';
+import {GALLERY_CONSTS, utils} from 'pro-gallery-lib';
 import { GalleryComponent } from '../galleryComponent';
 
 export default class ItemHover extends GalleryComponent {
@@ -30,18 +30,25 @@ export default class ItemHover extends GalleryComponent {
 
   render() {
     const {
-      shouldHover,
       imageDimensions,
       actions,
       idx,
+      itemWasHovered,
+      renderCustomInfo,
+      styleParams,
     } = this.props;
-    if (!shouldHover) {
-      return null;
-    }
     const hoverClass = this.getHoverClass();
     const { marginLeft, marginTop, width, height, ...restOfDimensions } =
       imageDimensions || {};
     //width and height will be taken from the gallery.scss and not be inline
+
+    const { hoveringBehaviour, overlayAnimation } = styleParams;
+    const { APPEARS } = GALLERY_CONSTS.infoBehaviourOnHover;
+    const { NO_EFFECT } = GALLERY_CONSTS.overlayAnimations;
+    //when there is a specific overlayAnimation, to support the animation we render the itemHover before any hover activity (see 'shouldHover()' in itemView).
+    //so in this specific case, the itemHover exists right away, but we do'nt want to render yet the hover-inner,
+    //the hover-inner will be rendered only after (at) the first hover an on, and not before.
+    const shouldRenderHoverInnerIfExist = hoveringBehaviour === APPEARS && overlayAnimation !== NO_EFFECT ? itemWasHovered : true;
 
     return (
       <div
@@ -56,9 +63,9 @@ export default class ItemHover extends GalleryComponent {
           onTouchStart={actions.handleItemMouseDown}
           onTouchEnd={actions.handleItemMouseUp}
         >
-          {this.props.renderCustomInfo ? (
+          {shouldRenderHoverInnerIfExist && renderCustomInfo ? (
             <div className="gallery-item-hover-inner">
-              {this.props.renderCustomInfo()}
+              {renderCustomInfo()}
             </div>) : null}
         </div>
       </div>